+++ authority-portal-backend/authority-portal-quarkus/src/main/kotlin/de/sovity/authorityportal/broker/services/api/filtering/CatalogFilterService.kt
@@ -21,11 +21,16 @@ import de.sovity.authorityportal.broker.dao.pages.catalog.CatalogQueryFields
 import de.sovity.authorityportal.broker.dao.pages.catalog.models.CatalogQueryFilter
 import de.sovity.authorityportal.broker.dao.pages.catalog.models.CatalogQuerySelectedFilterQuery
 import de.sovity.authorityportal.broker.dao.utils.JsonDeserializationUtils.read2dStringList
+import de.sovity.authorityportal.web.environment.CatalogDataspaceConfigService
+import de.sovity.authorityportal.web.environment.DeploymentEnvironmentService
 import jakarta.enterprise.context.ApplicationScoped
+import org.eclipse.microprofile.config.inject.ConfigProperty
 
 @ApplicationScoped
 class CatalogFilterService(
-    val catalogFilterAttributeDefinitionService: CatalogFilterAttributeDefinitionService
+    val catalogFilterAttributeDefinitionService: CatalogFilterAttributeDefinitionService,
+    val deploymentEnvironmentService: DeploymentEnvironmentService,
+    val catalogDataspaceConfigService: CatalogDataspaceConfigService
 ) {
 
     private val caseInsensitiveEmptyStringLast = Comparator<String> { s1, s2 ->
@@ -42,13 +47,13 @@ class CatalogFilterService(
          *
          * @return attribute definitions
          */
-        get() = listOf(
+        get() = listOfNotNull(
             catalogFilterAttributeDefinitionService.forField(
                 { fields: CatalogQueryFields -> fields.dataSourceAvailabilityLabel },
                 "dataSourceAvailability",
                 "Data Offer Type"
             ),
-            catalogFilterAttributeDefinitionService.buildDataSpaceFilter(),
+            catalogFilterAttributeDefinitionService.buildDataSpaceFilter().takeIf { this.shouldSupportMultipleDataspaces() },
             catalogFilterAttributeDefinitionService.forField(
                 { fields: CatalogQueryFields -> fields.dataOfferTable.DATA_CATEGORY },
                 "dataCategory",
@@ -165,5 +170,14 @@ class CatalogFilterService(
         return cnfFilterValue.selectedAttributeValues
             .filter { it.selectedIds.isNotEmpty() }
             .associate { it.id to it.selectedIds }
+    }
+
+    private fun shouldSupportMultipleDataspaces(): Boolean {
+        deploymentEnvironmentService.findAll().forEach {
+            if (catalogDataspaceConfigService.forEnvironment(it.key).namesByConnectorId.isNotEmpty()) {
+                return true
+            }
+        }
+        return false
     }
 }